Samples: Initial commit for bios-side samples
This patch adds the bios-side samples for OMAP4 M3 cores, and
tests the basic messaging, as well as resource allocations and
constraints framework. The samples also allow to test the power
management by including IpcPower module in the config files.

The build of these samples leverage the platform and
common config files.

Signed-off-by: G Anthony <a0783926@ti.com>
Signed-off-by: Mark Grosen <mgrosen@ti.com>
Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>