Utils: Update the firmware to carry section physical addresses
The utils.mmap and rsc_table.h.mmap versions (firmware v2 format with
additional 64-bit field for physical addresses) have been made the
mainline version. The BIOS image version has been bumped up to reflect
the new firmware section changes.

The rsc_table has also been adjusted for the resource table order. The
IPU_MEM_IPC, which contains the vring buffers, needs to be first if
using dynamic carveout. This is due to the current restrictions on the
Linux-side as to how the vring buffers are set up. The IPU_MEM_DATA is
moved down so that it would not affect the behavior if this carveout
section is made dynamic in the future too.

The flags field in resource_type is reverted back to reserved, and the
current IPU_MEMORY_FLAGS are removed since they provide no value and
the settings are already configured by default on the A9-side.

Signed-off-by: Suman Anna <s-anna@ti.com>