init: temporarily delay tasks init to avoid race
Yeah, this is horrible.

Without it, we might try to send a message before buffers are
available in the ring, and then the code just bluntly fails.

The real fix is to pause initialization until buffers are available,
or allow users to get notified when a buffer is available after
they failed getting one.

Signed-off-by: Ohad Ben-Cohen <ohad@wizery.com>