rsc_table: express vring sizes in number of supported buffers
The sizes of the vrings should be expressed in number of supported
buffers (which should be a power of 2).

Currently we support 256 buffers, but that can be easily changed now
solely on the sysbios side: the Linux side will just grab the number
from the resource table and build an appropriate vring.

Signed-off-by: Ohad Ben-Cohen <ohad@wizery.com>