Build: Set proper entry point for DSP
The entry point is set by default to point to c_int00 for
all targets. However, DSP has an additional alignment
requirement for the boot address to be aligned on a 1K
boundary. The ti_sysbios_family_c64p_Hwi0 is already
aligned by SYS/BIOS on a 1K boundary. This is the address
of vector 0 and branches off to c_int00.

This patch sets this symbol as the entry point for DSP
through a linker option for all sysbios-rpmsg builds,
thereby satisfying the alignment requirement.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>