rpmsgcio: hook up RpmsgCio to the virtio console device
This way sys/bios stdin/stdout will be routed to the Linux virtio
console.

This is just a preliminary implementation, mainly for demonstration
sake. known bugs that have to be fixed: vrings protection against
concurrent access, fix delicate races that might exist in the code, ..

IOW - this is not a product quality. just a prototype.

While at it, add a 'dump' command which Linux can send, which triggers
a dumping of accumulated early logs.

Signed-off-by: Ohad Ben-Cohen <ohad@wizery.com>